Restructure masher for bazel.
