--- package.json
@@ -1,14 +1,0 @@
-{
-  "name": "sembase",
-  "workspaces": [
-    "projects/*"
-  ],
-  "scripts": {
-    "dep-update": "git submodule update --init",
-    "dep-catch3": "cd lib/catch3 && npx bazelisk build //...",
-    "postinstall": "npm run dep-update; npm run dep-catch3"
-  },
-  "dependencies": {
-    "@bazel/bazelisk": "^1.19.0"
-  }
-}
