--- projects/masher/Makefile
@@ -1,10 +1,0 @@
-BOARD_TAG    = leonardo
-BOARDS_TXT   = /usr/share/arduino/hardware/arduino/avr/boards.txt
-MONITOR_PORT = /dev/ttyACM0
-AVR_TOOLS_DIR = /usr
-
-ARDUINO_LIBS = HID Keyboard
-USER_LIB_PATH = $(HOME)/Arduino/libraries
-ARDUINO_LIB_PATH = /usr/share/arduino/hardware/arduino/avr/libraries
-
-include ../../lib/Arduino-Makefile/Arduino.mk
