--- projects/masher/config.h
@@ -1,9 +1,0 @@
-#ifndef MASH_CFG_H_6324
-#define MASH_CFG_H_6324
-
-namespace cfg
-{
-    int mashDelay = 100;
-}
-
-#endif
