+++ projects/masher/lib/gpio/BUILD
@@ -1,0 +1,11 @@
+load(
+    "@platformio_rules//platformio:platformio.bzl",
+    "platformio_library",
+)
+
+platformio_library(
+    name = "gpio",
+    hdr = "gpio.h",
+    src = "gpio.cpp",
+    visibility = ["//visibility:public"],
+)
