+++ projects/masher/lib/key/BUILD
@@ -1,0 +1,11 @@
+load(
+    "@platformio_rules//platformio:platformio.bzl",
+    "platformio_library",
+)
+
+platformio_library(
+    name = "key",
+    hdr = "key.h",
+    src = "key.cpp",
+    visibility = ["//visibility:public"],
+)
